// examples/polls_app/frontend/src/guitar/models/Question.ts
/**
 * Question Model
 * Auto-generated by Django Guitar (placeholder)
 */

import { QuerySet } from '../client'

export interface Question {
  id: number
  question_text: string
  pub_date: string
}

export interface QuestionCreate {
  question_text: string
  pub_date?: string
}

export interface QuestionUpdate {
  question_text?: string
  pub_date?: string
}

class QuestionQuerySet extends QuerySet<Question, QuestionCreate, QuestionUpdate> {
  constructor() {
    super('/question')
  }
}

export const Question = {
  objects: new QuestionQuerySet(),
}