// examples/polls_app/frontend/src/guitar/models/Choice.ts
/**
 * Choice Model
 * Auto-generated by Django Guitar (placeholder)
 */

import { QuerySet } from '../client'

export interface Choice {
  id: number
  question_id: number
  choice_text: string
  votes: number
}

export interface ChoiceCreate {
  question_id: number
  choice_text: string
}

export interface ChoiceUpdate {
  choice_text?: string
}

class ChoiceQuerySet extends QuerySet<Choice, ChoiceCreate, ChoiceUpdate> {
  constructor() {
    super('/choice')
  }
}

export const Choice = {
  objects: new ChoiceQuerySet(),
}